+++ src/i18n.js
@@ -1,203 +1,4 @@
-const translations = {
-  en: {},
-  he: {
-    'What type is this photo?': 'איזה סוג תמונה זו?',
-    'Drivers': 'נהגים',
-    'Constructors': 'קבוצות',
-    'Current Team': 'קבוצה נוכחית',
-    'Sorry, I only support text and image messages.': 'מצטער, אני תומך רק בהודעות טקסט ותמונות.',
-    'Photo labeled as {TYPE}. Wait for extracted JSON data...': 'התמונה סומנה כ{TYPE}. נא להמתין לחילוץ נתונים...',
-    'An error occurred while extracting data from the photo.': 'אירעה שגיאה בעת חילוץ הנתונים מהתמונה.',
-    'Selected chip: {CHIP}.': 'צ\'יפ נבחר: {CHIP}.',
-    'Note: best team calculation was deleted.\nrerun {CMD} command to recalculate best teams.': 'לתשומת לבך: החישוב נמחק.\nהפעל את הפקודה {CMD} מחדש לחישוב.',
-    'which chip do you want to use?': 'איזה צ\'יפ תרצה להשתמש?',
-    'Extra DRS': 'DRS נוסף',
-    'Limitless': 'ללא הגבלה',
-    'Wildcard': 'ווילדקארד',
-    'Without Chip': 'ללא צ\'יפ',
-    'Sorry, only admins can use this command.': 'מצטער, רק מנהלים יכולים להשתמש בפקודה זו.',
-    'Simulation data fetched and cached successfully.': 'נתוני הסימולציה נטענו ונשמרו בהצלחה.',
-    'Failed to load simulation data: {ERROR}': 'נכשל לטעון נתוני סימולציה: {ERROR}',
-    'Cache has been reset for your chat.': 'המטמון אופס עבור הצ\'אט שלך.',
-    'Next race information is currently unavailable.': 'מידע על המרוץ הבא אינו זמין כעת.',
-    'Missing cached data. Please send images or JSON data for drivers, constructors, and current team first.':
-      'נתוני מטמון חסרים. אנא שלח תמונות או קבצי JSON של נהגים, קבוצות וקבוצה נוכחית קודם.',
-    'Please send a number to get the required changes to that team.':
-      'אנא שלח מספר לקבלת השינויים הדרושים לקבוצה הזו.',
-    'You currently have data in your cache. To use data from a simulation, please run {CMD} first.':
-      'יש לך נתונים במטמון. להפעלת נתוני סימולציה הפעל קודם את {CMD}.',
-    'No simulation data is currently loaded. Please use {CMD} to load simulation data.':
-      'אין נתוני סימולציה טעונים. השתמש ב-{CMD} כדי לטעון נתונים.',
-    'Last updated: {TIME}': 'עודכן לאחרונה: {TIME}',
-    'Current simulation: {NAME}\n{UPDATE}': 'סימולציה נוכחית: {NAME}\n{UPDATE}',
-    '💡 Tip: If the simulation seems outdated, you can run {CMD} to update the current simulation.':
-      '💡 טיפ: אם הסימולציה נראית מיושנת, ניתן להריץ את {CMD} לעדכון.',
-    'Sorry, only admins can access billing statistics.': 'מצטער, רק מנהלים יכולים לצפות בסטטיסטיקת החיוב.',
-    '❌ Error fetching billing statistics: {ERROR}\n\nPlease check your Azure configuration and permissions.':
-      '❌ שגיאה בקבלת סטטיסטיקת חיוב: {ERROR}\n\nאנא בדוק את ההגדרות וההרשאות של Azure.',
-    'No billing data available for this period.': 'אין נתוני חיוב לתקופה זו.',
-    'Azure Billing Statistics': 'סטטיסטיקת חיוב Azure',
-    'Current Month': 'חודש נוכחי',
-    'Previous Month': 'חודש קודם',
-    'Month-over-Month Comparison:': 'השוואת חודש לחודש:',
-    'Increase': 'עלייה',
-    'Decrease': 'ירידה',
-    'No change': 'ללא שינוי',
-    'Unknown menu action': 'פעולת תפריט לא ידועה',
-    '🎯 *F1 Fantasy Bot Menu*\n\nChoose a category:': '🎯 *תפריט הבוט*\n\nבחר קטגוריה:',
-    'Tip:': 'טיפ:',
-    'Use {CMD} for quick text-based help': 'השתמש ב-{CMD} לעזרה מהירה',
-    '❓ Help': '❓ עזרה',
-    'Show this help message.': 'הצג הודעת עזרה זו.',
-    '⬅️ Back to Main Menu': '⬅️ חזרה לתפריט הראשי',
-    'Choose a command:': 'בחר פקודה:',
-    'Executing {CMD}...': 'מריץ את {CMD}...',
-    'Error executing command': 'שגיאה בהרצת פקודה',
-    'Command not found': 'הפקודה לא נמצאה',
-    'Showing help...': 'מציג עזרה...',
-    'Error showing help': 'שגיאה בהצגת עזרה',
-    'Drivers cache is empty. Please send drivers image or valid JSON data.':
-      'מטמון הנהגים ריק. אנא שלח תמונת נהגים או נתוני JSON תקינים.',
-    'Selected Chip: {CHIP}': 'צ\'יפ שנבחר: {CHIP}',
-    'No chip selected.': 'לא נבחר צ\'יפ.',
-    'Team {NUM} Required Changes:': 'שינויים דרושים לקבוצה {NUM}:',
-    'Drivers To Add': 'נהגים להוספה',
-    'Drivers To Remove': 'נהגים להסרה',
-    'Constructors To Add': 'קבוצות להוספה',
-    'Constructors To Remove': 'קבוצות להסרה',
-    'Extra DRS Driver': 'נהג DRS נוסף',
-    'New ': 'חדש ',
-    'DRS Driver': 'נהג DRS',
-    'Chip To Activate': 'צ\'יפ להפעלה',
-    'Team {NUM} Info:': 'מידע לקבוצה {NUM}:',
-    'Projected Points': 'נקודות צפויות',
-    'Expected Price Change': 'שינוי מחיר צפוי',
-    'Total Price': 'מחיר כולל',
-    'Transfers Needed': 'מספר העברות נדרש',
-    'Penalty': 'קנס',
-    'Team {NUM} (Current Team)': 'קבוצה {NUM} (קבוצה נוכחית)',
-    'Team {NUM}': 'קבוצה {NUM}',
-    'Drivers & Constructors Total Price': 'מחיר כולל של נהגים וקבוצות',
-    'Cost Cap Remaining': 'יתרת תקציב',
-    'Total Budget': 'תקציב כולל',
-    'Expected Points': 'נקודות צפויות',
-    'Δ Points': 'Δ נקודות',
-    'Δ Price': 'Δ מחיר',
-    'Weather Forecast': 'תחזית מזג אוויר',
-    'Sprint Qualifying': 'מקצה דירוג ספרינט',
-    'Sprint': 'ספרינט',
-    'Qualifying': 'דירוג',
-    'Race': 'מרוץ',
-    'Next Race Information': 'מידע על המרוץ הבא',
-    'Race Name': 'שם המרוץ',
-    'Track': 'מסלול',
-    'Location': 'מיקום',
-    'Sprint Qualifying Date': 'תאריך דירוג ספרינט',
-    'Sprint Qualifying Time': 'שעת דירוג ספרינט',
-    'Sprint Date': 'תאריך ספרינט',
-    'Sprint Time': 'שעת ספרינט',
-    'Qualifying Date': 'תאריך דירוג',
-    'Qualifying Time': 'שעת דירוג',
-    'Race Date': 'תאריך מרוץ',
-    'Race Time': 'שעת מרוץ',
-    'Weekend Format': 'פורמט סוף שבוע',
-    'Historical Race Stats (Last Decade)': 'סטטיסטיקת מרוצים היסטורית (עשור אחרון)',
-    'Pole': 'פול',
-    'Winner': 'מנצח',
-    '2nd': 'מקום שני',
-    '3rd': 'מקום שלישי',
-    'Cars Finished': 'מכוניות שסיימו',
-    'Overtakes': 'עקיפות',
-    'Safety Cars': 'מכוניות בטיחות',
-    'Red Flags': 'דגלים אדומים',
-    'No historical data available for this track.': 'אין נתונים היסטוריים למסלול זה.',
-    'Track History': 'היסטוריית מסלול',
-    'F1 Fantasy Bot - Available Commands': 'פקודות זמינות - F1 Fantasy Bot',
-    'Other Messages': 'הודעות נוספות',
-    'Send an image (drivers, constructors, or current team screenshot) to automatically extract and cache the relevant data.':
-      'שלח תמונה (נהגים, קבוצות או צילום של הקבוצה הנוכחית) לחילוץ אוטומטי ושמירת הנתונים.',
-    'Send valid JSON data to update your drivers, constructors, and current team cache.':
-      'שלח נתוני JSON תקינים לעדכון המטמון של הנהגים, הקבוצות והקבוצה הנוכחית שלך.',
-    'Send a number (e.g., 1) to get the required changes to reach that team from your current team (after using {CMD}).':
-      'שלח מספר (לדוגמה, 1) כדי לקבל את השינויים הנדרשים להגיע לקבוצה זו מהקבוצה הנוכחית (לאחר שימוש ב-{CMD}).'
-    ,
-    'Unknown': 'לא ידוע',
-    'Invalid date': 'תאריך לא תקין'
-    ,
-    'Language changed to {LANG}.': 'השפה שונתה ל{LANG}.',
-    'Invalid language. Supported languages: {LANGS}':
-      'שפה לא תקינה. השפות הזמינות: {LANGS}',
-    'Usage: {CMD} <LANG>': 'שימוש: {CMD} <שפה>',
-    'Please select a language:': 'אנא בחר שפה:',
-    English: 'אנגלית',
-    Hebrew: 'עברית',
-    '🌐 Set Language': '🌐 הגדר שפה',
-    'Change bot language for this session': 'שנה את שפת הבוט להפעלה זו',
-    '❓ Help & Menu': '❓ עזרה ותפריט',
-    'Help and navigation commands': 'פקודות עזרה וניווט',
-    '🏎️ Team Management': '🏎️ ניהול קבוצה',
-    'Manage and optimize your F1 Fantasy team':
-      'ניהול ואופטימיזציה של קבוצת F1 Fantasy',
-    '📊 Analysis & Stats': '📊 ניתוח וסטטיסטיקה',
-    'View race information and performance data':
-      'צפה במידע על מרוצים ונתוני ביצועים',
-    '🔧 Utilities': '🔧 כלי עזר',
-    'Data management and maintenance tools': 'כלי ניהול ותחזוקה של נתונים',
-    '👤 Admin Commands': '👤 פקודות מנהל',
-    'Administrative tools and functions': 'כלי ותפקידי ניהול',
-    '📱 Menu': '📱 תפריט',
-    'Show interactive menu with all available commands.':
-      'הצג תפריט אינטראקטיבי עם כל הפקודות הזמינות.',
-    '🏆 Best Teams': '🏆 הקבוצות הטובות ביותר',
-    'Calculate and display the best possible teams based on your cached data':
-      'חשב והצג את הקבוצות הטובות ביותר על סמך הנתונים במטמון',
-    '👥 Current Team Info': '👥 מידע על הקבוצה הנוכחית',
-    'Calculate the current team info based on your cached data':
-      'חשב את מידע הקבוצה הנוכחית על סמך הנתונים במטמון',
-    '🎯 Chips Selection': "🎯 בחירת צ'יפ",
-    'Choose a chip to use for the current race': 'בחר צ\'יפ לשימוש במרוץ הנוכחי',
-    '🏁 Next Race Info': '🏁 מידע על המרוץ הבא',
-    'Get detailed information about the next F1 race':
-      'קבל מידע מפורט על המרוץ הבא',
-    '📈 Current Simulation': '📈 סימולציה נוכחית',
-    'Show the current simulation data and name':
-      'הצג את נתוני הסימולציה הנוכחית ואת שמה',
-    '📄 Print Cache': '📄 הדפסת מטמון',
-    'Show the currently cached drivers, constructors, and current team':
-      'הצג את הנהגים, הקבוצות והקבוצה הנוכחית במטמון',
-    '🗑️ Reset Cache': '🗑️ איפוס מטמון',
-    'Clear all cached data for this chat': 'נקה את כל הנתונים במטמון עבור צ\'אט זה',
-    '📋 Load Simulation': '📋 טעינת סימולציה',
-    'Load latest simulation data': 'טען את נתוני הסימולציה העדכניים ביותר',
-    '🔄 Trigger Scraping': '🔄 הפעל סריקה',
-    'Trigger web scraping for latest F1 Fantasy data':
-      'הפעל סריקת רשת לקבלת נתוני F1 Fantasy העדכניים',
-    '🤖 BotFather Commands': '🤖 פקודות BotFather',
-    'Get commands for BotFather setup': 'קבל פקודות להגדרת BotFather',
-    '💰 Billing Stats': '💰 סטטיסטיקת חיוב',
-    'Get Azure billing statistics for the current month':
-      'קבל סטטיסטיקת חיוב Azure לחודש הנוכחי',
-    'ℹ️ Version': 'ℹ️ גרסה',
-    'Show current deployed version': 'הצג את הגרסה המותקנת הנוכחית',
-    'Invalid JSON data. Please ensure it contains 20 drivers under "Drivers" property.':
-      'נתוני JSON אינם תקינים. ודא שהם מכילים 20 נהגים תחת "Drivers".',
-    'Invalid JSON data. Please ensure it contains 10 constructors under "Constructors" property.':
-      'נתוני JSON אינם תקינים. ודא שהם מכילים 10 קבוצות תחת "Constructors".',
-    'Invalid JSON data. Please ensure it contains the required properties under "CurrentTeam" property.':
-      'נתוני JSON אינם תקינים. ודא שהם מכילים את המאפיינים הנדרשים תחת "CurrentTeam".',
-    'Error: Scraping trigger URL is not configured.': 'שגיאה: כתובת ההפעלה לסריקה אינה מוגדרת.',
-    'Sorry, only admins can trigger scraping.': 'מצטער, רק מנהלים יכולים להפעיל סריקה.',
-    'Web scraping triggered successfully.': 'סריקת הרשת הופעלה בהצלחה.',
-    'Failed to trigger web scraping: {ERROR}': 'נכשל להפעיל סריקה: {ERROR}',
-    'Commit ID: {ID}\nCommit message: {MSG}\nLink: {LINK}':
-      'מזהה קומיט: {ID}\nהודעת קומיט: {MSG}\nקישור: {LINK}'
-  }
-};
-
-const LANGUAGE_NAME_KEYS = {
-  en: 'English',
-  he: 'Hebrew',
-};
+const { translations, LANGUAGE_NAME_KEYS } = require('./translations');
 const { languageCache } = require('./cache');
 let defaultLanguage = process.env.BOT_LANGUAGE || 'en';
